==> explore_figure/covid_validation_3/01_covid_h5_merge.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Matrix)
library(rhdf5)
library(tidyverse)
library(glue)
library(here)
library(hdf5r)
library(Seurat)
```

# COVID_BALF
```{r}
# C141
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339769_C141_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val1 <- CreateSeuratObject(data)
val1
```

```{r}
# C142
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339770_C142_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val2 <- CreateSeuratObject(data)
val2
```

```{r}
# C143
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339771_C143_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val3 <- CreateSeuratObject(data)
val3
```

```{r}
# C144
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339772_C144_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val4 <- CreateSeuratObject(data)
val4
```

```{r}
# C145
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339773_C145_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val5 <- CreateSeuratObject(data)
val5
```

```{r}
# C146
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339774_C146_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val6 <- CreateSeuratObject(data)
val6
```

```{r}
# C51
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475048_C51_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val7 <- CreateSeuratObject(data)
val7
```

```{r}
# C52
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475049_C52_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val8 <- CreateSeuratObject(data)
val8
```

```{r}
# C100
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475050_C100_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val9 <- CreateSeuratObject(data)
val9
```

```{r}
# C148
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475051_C148_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val10 <- CreateSeuratObject(data)
val10
```

```{r}
# C149
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475052_C149_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val11 <- CreateSeuratObject(data)
val11
```

```{r}
# C152
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475053_C152_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val12 <- CreateSeuratObject(data)
val12
```






# COVID_lung 

```{r}
# C166
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516279_C166_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj1 <- CreateSeuratObject(data)
obj1
```

```{r}
# C168
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516280_C168_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj2 <- CreateSeuratObject(data)
obj2
```

```{r}
# C170
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516281_C170_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj3 <- CreateSeuratObject(data)
obj3
```

```{r}
# C172
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516282_C172_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj4 <- CreateSeuratObject(data)
obj4
```









```{r}
# merge SeuratObject - COVID BALF and lung data 
covid_merge <- merge(val1, y = c(val2, val3, val4, val5, val6, val7, val8, val9, val10, val11, val12, obj1, obj2, obj3, obj4), add.cell.ids = c('val1', 'val2', 'val3', 'val4', 'val5', 'val6','val7', 'val8', 'val9', 'val10', 'val11', 'val12', 'obj1', 'obj2', 'obj3', 'obj4'), project = 'scRNA_IPF_data')
covid_merge
```


```{r}
unique(sapply(X = strsplit(colnames(covid_merge), split = "_"), FUN = "[", 1))
```


```{r}
saveRDS(covid_val, file = "01_COVID_merge_h5_to_SeuratObject_merge.RDS")
```
